requirements.txt
azure-storage-blob==12.10.0
sqlalchemy
